requirements.txt
# Telegram
telethon==1.34.0
cryptg==0.4.0

# Database
supabase==2.3.4
python-dotenv==1.0.0

# Utilities
aiohttp==3.9.1
asyncio==3.4.3
python-dateutil==2.8.2

# Logging & Monitoring
colorlog==6.8.0
tabulate==0.9.0

# Background tasks
schedule==1.2.0
